Fix tests, lints, resolve dependency cycle
