+++ apps/website/src/components/Nav/_DesktopNavLinks.tsx
@@ -1,24 +1,24 @@
 import { NavLinks } from './_NavLinks';
-import { ExpandedSectionsState } from './Nav';
+import { ExpandedSectionsState } from './utils';
 
 export const DesktopNavLinks: React.FC<{
-    expandedSections: ExpandedSectionsState;
-    updateExpandedSections: (updates: Partial<ExpandedSectionsState>) => void;
-    courses: { title: string; url: string; isNew?: boolean }[];
-    isScrolled: boolean;
-  }> = ({
-    expandedSections,
-    updateExpandedSections,
-    courses,
-    isScrolled,
-  }) => {
-    return (
-      <NavLinks
-        className="desktop-nav-links__nav-links hidden lg:flex"
-        expandedSections={expandedSections}
-        updateExpandedSections={updateExpandedSections}
-        courses={courses}
-        isScrolled={isScrolled}
-      />
-    );
-  };
+  courses: { title: string; url: string; isNew?: boolean }[];
+  expandedSections: ExpandedSectionsState;
+  updateExpandedSections: (updates: Partial<ExpandedSectionsState>) => void;
+  isScrolled: boolean;
+}> = ({
+  courses,
+  expandedSections,
+  updateExpandedSections,
+  isScrolled,
+}) => {
+  return (
+    <NavLinks
+      courses={courses}
+      expandedSections={expandedSections}
+      updateExpandedSections={updateExpandedSections}
+      isScrolled={isScrolled}
+      className="hidden lg:flex"
+    />
+  );
+};
